requirements.txt
accelerate==1.7.0
backoff==2.2.1
bitsandbytes==0.45.5
datasets==3.6.0
fire==0.7.0
google-genai==1.53.0
openai==1.99.1
packaging==25.0
pandas==2.3.1
pydantic==2.12
PyYAML==6.0.2
scikit-learn==1.5.2
setuptools==80.9.0
torch==2.9.0
tqdm==4.67.1
transformers==4.57.3
unsloth==2025.5.9
vllm==0.11.2
seaborn==0.13.2
matplotlib==3.10.8
seaborn==0.13.2
numpy==2.4.4
scipy==1.17.1